requirements.txt
PyQt6==6.7.0
Pillow==10.3.0
reportlab==4.2.0
pyinstaller==6.3.0
